Added bunch of things

+++ karp_rabin_hash.h
@@ -9,11 +9,11 @@
 //hash generator type
 typedef struct _hash_gen_t {
     int kgram_len;
-    int prime;
+    long prime;
     int h;
     hash_t curr_hash;
 } hash_gen_t;
 
-hash_t create_hash_generator(hash_gen_t** generator, int kgram_len, int prime, char* first_kgram);
+hash_t create_hash_generator(hash_gen_t** generator, int kgram_len, long prime, char* first_kgram);
 
 hash_t generate_next_hash(hash_gen_t* generator, char* new_kgram);
